Sales Dashboard of a UK Toy company
It shows the sales of the products across the countries. The Quantity sold was identified to be 513 while the sum of sales price was indicated to be 240K. from the analysis it was made known that southwest has the highest sales. A high sale in October with yellow as the  most ordered product. United states got the highest demand. It is recommended that a market survey is needed for countries with low sales and the reason for getting the yellow products more than others should be checked.