Meriskill Company Sales Data
The Dataset indicated sales information which shows the columns description of sales as well as some of the unique identifier of the sales transactions. The table displays columns such as
•	ORDER ID
•	PRODUCT
•	QUANTITY ORDERED
•	PRICE EACH
•	ORDER DATE
•	PURCHASE ADDRESS
•	MONTH
•	SALE
•	CITY
•	HOUR
From the visualization obtained, the visualization displays the Total revenue obtained by the organization, Quantity sold, Cost of products (Production costs), Profit margin, counts of products and Number of orders Made. 
From the Key indicators (KPI), a total Revenue of 34,492K was made with a quantity sale of the products to be 209K. Furthermore, the cost of production was 34,289K and the profit margin was 58.8%. In the counts of products, a total number of 19 products was recalled and the numbers of order made by customers was 178K. Hence, this simply shows that XYZ Technologies deals in 19 different products. 
According to the visualization obtained, XYZ technologies recalled a high sale of their products in December and the top 5 cities was analyzed which are Boston, Dallas, Los Angeles, New York and San Fransisco. Moreover, on weekly sales of products, Tuesdays recalled the highest sales of the products being sold by the company. Observing the top 5 products by their sales counts, MacBook recall the highest, follow by iPhone, Laptop and the rest. In the Top five best selling products, AAA batteries, AA batteries, USB charger, Lightning Charger and was observed to be the highest and wired Headphones were some of the products in this category. 
Recommendations 
The recommendations put in place by the analyst include the following:  
•	The need for the organization to do market research on reasons why their products are selling more in the stated cities with the highest sales. 
•	Tuesdays was observed to have the highest sales. The need for more market strategic planning on days with low sales so as to boost the market sales along all days of the week. 
•	MacBook was noted to be the highest among all the products in term of sales counts. More research is needed to be done on why people prefer the particular products more than the others.