B&G Supermarket Pizza Sales trend
The Pizza Sales shows trend in the sales of Pizza by the organization. The Data set contains columns that relate to the transactions and pattern of sales in Pizza in the organization. Examining the key performance Indicators (KPI), the Total Revenue was noted, Average Revenue, Total Order made and the Quantity Sold. This was used to form explainable insights into the trends and patterns of sales in the organization. The total revenue generated by the organization was 817.86K while the total quantity of Pizza sold was noted to be 50K, hence showing that the quantity of products sold was multiplied by the sales to provide the revenue generated by the organization. 
On the insights built via the visualization, it was observed that in the daily sales trend, Friday gets the highest sales. This could be as a result of the need for people or customers to go out and enjoy the weekend and hangouts, hence bringing more sales to the organization.  Moreover, the Monthly sales recalled a high sale of 4,392 pizzas in July which could be as a result of the marketing strategy put in place by the organization as the month generated the highest of all the months considered in the sales. Furthermore, on the Pizza category, the classic made the highest sales while the large pizza size brought in a high turnover. The sales of these categories could be because of the taste or the packaging the organization provided to it. 
Therefore, some recommendations were put in place to aid more sales and good market placement for the organization. The recommendations include the following:  
1.	The need for the organization to ensure that more productions are done on Friday for more sales as it is the period of the week when people want to wine and dine to the weekend. 
2.	On the monthly sales, it could be seen that October recalled the lowest sales of 3883 in the organization. Therefore, it is recommended that the organization need to check the reason for low sales in October and high sales in July, using this reason to change the strategy of production in October for better sales in months with low sales. 
3.	The reason why there are a high demand for the classic pizza category should be cross-checked through a deep market survey. Hence the finding should be adopted in dealing with other pizza category in the organization. 
4.	Better value for money could be the reason why people preferred the large sizes pizza. So, the organization need to investigate more into this and make adjustment when need be to the medium and small ones.