Nwachukwu Technologies Sales Dashboard
The dashboard analyses the sales information of Mr. Nwachukwu Technologies company by showing the total revenue accrue by the company and the units sold in the period of the years and the locations. Insights were built around the sales information using the pivot table which summaries the company revenue based on the locations or regions and the years. The sum of 73,031,997,790 were generated as the sum of revenue for the company and in which Abia state recalled the highest sales and states that generated the highest revenue for the company. Furthermore, on the revenue by region, the North east generated the highest revenue for the company which was followed by the south west, hence the North central is the least. The revenue by rank shows an average performance was given to the company for the years. The trend on the revenue by months shows that the highest revenue was recalled by the company in May. 
It is recommended that a good marketing strategy should be adopted by the marketing department on region (North central) with poor performance. As this on the long run will boost the sales and revenue.